requirements.txt
pyTelegramBotAPI
aiohttp
aiodns
ujson
tortoise-orm
